icn_mesh

Module metrics

Statics

JOB_PROCESS_TIME

Records the time from job assignment to receipt processing in seconds.

PENDING_JOBS_GAUGE

Tracks the number of jobs currently waiting in the runtime queue.

SCHEDULE_MESH_JOB_CALLS

Counts calls to schedule_mesh_job.

SELECT_EXECUTOR_CALLS

Counts calls to select_executor.