icn_mesh

Module metrics

Statics

BIDS_RECEIVED_TOTAL

Counts the total number of bids received across all jobs.

JOBS_ASSIGNED_TOTAL

Counts the total number of jobs that have been assigned to executors.

JOBS_BIDDING_GAUGE

Tracks the number of jobs currently in the bidding phase.

JOBS_COMPLETED_TOTAL

Counts the total number of jobs that have been completed successfully.

JOBS_EXECUTING_GAUGE

Tracks the number of jobs currently being executed.

JOBS_FAILED_TOTAL

Counts the total number of jobs that have failed.

JOBS_SUBMITTED_TOTAL

Counts the total number of jobs submitted to the mesh.

JOB_ASSIGNMENT_TIME

Records the time from job submission to assignment in seconds.

JOB_COMPLETION_TIME

Records the time from job submission to completion in seconds.

JOB_PROCESS_TIME

Records the time from job assignment to receipt processing in seconds.

PENDING_JOBS_GAUGE

Tracks the number of jobs currently waiting in the runtime queue.

SCHEDULE_MESH_JOB_CALLS

Counts calls to schedule_mesh_job.

SELECT_EXECUTOR_CALLS

Counts calls to select_executor.